OPTIMISATION: création de fichiers  pour chaque graphique
