+++ frontend/src/pages/ChartsCO2.jsx
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import options from "../options/options_CO2";
 import { Line } from "react-chartjs-2";
 import {
   Chart as ChartJS,
@@ -23,75 +24,36 @@ ChartJS.register(
 // constantes de données, api distante et copie locale
 const apiOnline = "https://global-warming.org/api/co2-api";
 const apiLocale = "./datasets/co2.json";
+const CO2 = "CO\u2082";
 // fin des constantes de données
 
-// constante de titrage du graphique
-const options = {
-  responsive: true,
-  scales: {
-    x: {
-      // configuration de l'axe horizontal
-      // ticks: {
-      //   // For a category axis, the val is the index so the lookup via getLabelForValue is needed
-      //   callback: function (val, index) {
-      //     // affichage d'une année, une colonne sur deux.
-      //     return index % 2 === 0 ? this.getLabelForValue(val) : "";
-      //   },
-      // },
-      display: true,
-      title: {
-        display: false,
-        text: "Année",
-      },
+/**
+ *  réglages des axes du tableau
+ */
+const dataModel = {
+  labels: [],
+  datasets: [
+    {
+      label: "Valeurs quotidiennes (cycle)", // titre du premier graphique
+      data: [],
+      pointStyle: "dash",
+      radius: 0.1,
+      borderColor: "rgba(136, 78, 160,0.8)",
+      backgroundColor: "rgba(136, 78, 160, 0.8)",
     },
-    y: {
-      // configuration de l'axe vertical
-      ticks: {
-        callback: function (value) {
-          //  callback: function (value, index, ticks) {
-          return `${value}`;
-        },
-      },
-      display: true,
-      title: {
-        // titre de l'axe vertical
-        display: true,
-        text: `Concentration de CO\u2082 en ppm`,
-      },
+    {
+      label: "Tendance moyenne (évolution)", // titre du second graphique
+      data: [],
+      pointStyle: "dash",
+      borderColor: "rgba(52,73,94, 0.8)", // "rgb(93, 109, 126)",
+      backgroundColor: "rgba(52, 73, 94, 0.8)",
     },
-  },
-  plugins: {
-    legend: {
-      position: "top",
-    },
-    title: {
-      display: true,
-      text: ``, // titre par défaut, remplacé via prepaConfig
-    },
-  },
+  ],
 };
 
 function ChartsCO2() {
-  const [dataSet, setDataSet] = useState({
-    labels: [],
-    datasets: [
-      {
-        label: "Valeurs quotidiennes (cycle)", // titre du premier graphique
-        data: [],
-        pointStyle: "dash",
-        radius: 0.1,
-        borderColor: "rgba(136, 78, 160,0.8)",
-        backgroundColor: "rgba(136, 78, 160, 0.8)",
-      },
-      {
-        label: "Tendance moyenne (évolution)", // titre du second graphique
-        data: [],
-        pointStyle: "dash",
-        borderColor: "rgba(52,73,94, 0.8)", // "rgb(93, 109, 126)",
-        backgroundColor: "rgba(52, 73, 94, 0.8)",
-      },
-    ],
-  });
+  const [dataSet, setDataSet] = useState(dataModel);
+
   const prepaConfig = (data) => {
     options.plugins.title.text = `Concentration de Dioxyde de Carbone dans l'atmosphère (de ${
       data[0].year
@@ -123,7 +85,6 @@ function ChartsCO2() {
   useEffect(() => {
     getStaticData(apiOnline, getStaticData);
   }, []);
-  const CO2 = "CO\u2082";
   return (
     // affichage du composant graphique
     <div className="graph_render_dd">
