+++ frontend/src/pages/ChartsTemp.jsx
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import options from "../options/options_temperature";
 import { Line } from "react-chartjs-2";
 import {
   Chart as ChartJS,
@@ -25,67 +26,26 @@ const apiOnline = "https://global-warming.org/api/temperature-api";
 const apiLocale = "./datasets/temperature.json";
 // fin des constantes de données
 
-// constante de titrage du graphique
-const options = {
-  responsive: true,
-  scales: {
-    x: {
-      // configuration de l'axe horizontal
-      ticks: {
-        // For a category axis, the val is the index so the lookup via getLabelForValue is needed
-        callback: function (val, index) {
-          // affichage d'une date arrondie à l'année, une colonne sur deux.
-          return index % 2 === 0 ? Math.round(this.getLabelForValue(val)) : "";
-        },
-      },
-      display: true,
-      title: {
-        display: false,
-        text: "Année",
-      },
+/**
+ *  réglages des axes du tableau
+ */
+const dataModel = {
+  labels: [],
+  datasets: [
+    {
+      label: "Température", // aka Station dans le set de données
+      data: [],
+      borderWidth: 1,
+      radius: 1,
+      borderColor: "rgba(255, 99, 132, 0.5)",
+      backgroundColor: "rgba(255, 99, 132, 0.8)",
+      color: "rgb(1, 99, 132)",
     },
-    y: {
-      // configuration de l'axe vertical
-      ticks: {
-        callback: function (value) {
-          // callback: function (value, index, ticks) {
-          return `${value} °C`;
-        },
-      },
-      display: true,
-      title: {
-        // titre de l'axe vertical
-        display: true,
-        text: "Température moyenne en °C",
-      },
-    },
-  },
-  plugins: {
-    legend: {
-      position: "top",
-    },
-    title: {
-      display: true,
-      text: `Évolution de la température moyenne`, // titre du graphique complet (date in:out)
-    },
-  },
+  ],
 };
 
 function ChartsTemp() {
-  const [dataSet, setDataSet] = useState({
-    labels: [],
-    datasets: [
-      {
-        label: "Température", // aka Station dans le set de données
-        data: [],
-        borderWidth: 1,
-        radius: 1,
-        borderColor: "rgba(255, 99, 132, 0.5)",
-        backgroundColor: "rgba(255, 99, 132, 0.8)",
-        color: "rgb(1, 99, 132)",
-      },
-    ],
-  });
+  const [dataSet, setDataSet] = useState(dataModel);
   const prepaConfig = (data) => {
     options.plugins.title.text = `Estimation de l'évolution de la température moyenne au sol (de ${Math.round(
       data[0].time
